refactor(daemon): split remote backend into parsers/state/watch modules
Reduces remote-backend.ts from 695 → 292 lines and brings each new file under the 400-line soft limit, satisfying the complexity ratchet. parsers, lifecycle state helpers (emit/settleTerminal/markRemoteExited), and the watch async generator each live in their own file. Behavior unchanged; all 15 unit tests still pass.
